refactor: split parse_max_chars's tests out of answer_assist.rs for r8
Merging main's real retry/cap-window machinery into this branch's own
maxChars addition pushed the module to 1482 lines, over the 1400-line
hard cap. Moved the already-isolated parse_max_chars_tests module into
its own file, the same way answer_assist_tests.rs already is. Pure
relocation, no assertion changed.
